flask_model/requirements.txt
flask
transformers
torch
Pillow
requests
